# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Blog post",
    "description": "Blog post management API",
    "termsOfService": "",
    "contact": {
      "name": "M2M Team",
      "email": "team@mind2minds.com",
      "url": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": ""
    },
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "paths": {
...
    "/api/medias": {
      "get": {
        "tags": ["Medias"],
        "description": "Get all medias",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/token"
            },
            "required": "true",
            "description": "User login token"
          }
        ],
        "responses": {
          "200": {
            "description": "List of all medias",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadMedia"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "status": "error",
                  "message": "Invalid token"
                }
              }
            }
          }
        }
      }
    },
    "/api/medias/{imageName}": {
      "post": {
        "tags": ["Medias"],
        "description": "Upload Media",
        "operationId": "UploadMedia",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/token"
            },
            "required": "true",
            "description": "User login token"
          },
          {
            "name": "id",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadMedia"
              }
            }
          },
          "required": "true"
        },
        "responses": {
          "200": {
            "description": "Media uploaded successfully"
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "status": "error",
                  "message": "Media not uploaded successfully"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Medias"],
        "description": "Delete Media",
        "operationId": "deleteMedia",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/token"
            },
            "required": "true",
            "description": "User login token"
          },
          {
            "name": "id",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media deleted successfully"
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "status": "error",
                  "message": "Media doesn't exist"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "id": {
        "type": "integer",
        "description": "unique identification number",
        "example": "1234"
      },
...
      "name": {
        "type": "string"
      },
...
      "type": {
        "type": "string"
      },
      "path": {
        "type": "string"
      },
      "mediaType": {
        "type": "string"
      },
      "mediaUrl": {
        "type": "string"
      },
...
      "UploadMedia": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/name"
          },
          "type": {
            "$ref": "#/components/schemas/type"
          },
          "path": {
            "$ref": "#/components/schemas/path"
          },
          "mediaType": {
            "$ref": "#/components/schemas/mediaType"
          },
          "mediaUrl": {
            "$ref": "#/components/schemas/mediaUrl"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "msg": {
            "type": "string"
          }
        }
      }
    }
  }
}